pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "omni_engineer"
version = "0.1.0"
description = "A self-improving assistant framework with tool creation capabilities"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [{ name = "Your Name", email = "your.email@example.com" }]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "anthropic",
    "python-dotenv",
    "rich",
    "requests",
    "beautifulsoup4",
    "validators>=0.34.0",
    "PyAutoGUI",
    "Pillow",
    "prompt-toolkit",
    "matplotlib>=3.9.2",
    "flask>=3.0.3",
    "werkzeug>=3.1.2",
    "markdownify>=0.14.1",
    "protego>=0.3.1",
    "readability-lxml>=0.8.1",
    "e2b-code-interpreter>=1.0.3",
    "aiohttp>=3.9.3",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "black",
    "ruff",
    "mypy",
    "pre-commit",
    "types-requests",
]

[project.urls]
Homepage = "https://github.com/yourusername/omni-engineer"
Repository = "https://github.com/yourusername/omni-engineer"
Documentation = "https://github.com/yourusername/omni-engineer#readme"
"Bug Tracker" = "https://github.com/yourusername/omni-engineer/issues"

[tool.hatch]
version.path = "omni_engineer/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["omni_core", "tools"]

[tool.ruff]
line-length = 88
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "RUF", # ruff-specific rules
]
ignore = []

[tool.ruff.isort]
known-first-party = ["omni_engineer"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.black]
line-length = 88
target-version = ["py39"]
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = ["wikipedia.*", "pyautogui.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=omni_engineer --cov-report=term-missing"
testpaths = ["tests"]

[tool.uv.workspace]
members = ["testfolder"]

[project.scripts]
omni_engineer = "omni_engineer.omni_engineer:main"
